MP3 project
My eye-controlled MP3 desktop application is a cutting-edge software that is currently in the development stage. The app is being coded using customtkinter and Python, and it is designed to provide a novel way of controlling MP3 playback. The application will use a dedicated eye-tracking device to capture the user's eye movements and recognize when the user blinks. When the user blinks, the application will play the next song in the playlist